fix(ng-dev): uncommitted changes check returning false-positives (#416)
In some cases the hasUncommittedChanges check returns false-positives
that occur due to us not refreshing & cleaning up the index before diffing
the index. The index needs a cleanup/refresh when a file is simply touched/modified with the actual content not changing. i.e. just the last modified time got updated

More details: https://stackoverflow.com/questions/34807971/why-does-git-diff-index-head-result-change-for-touched-files-after-git-diff-or-g/34808299#34808299