build: automatically set proper module name for first-party module resolution
We should set the module name based on the NPM output (like we do in other
repositories) so that our own files could resolve ng-dev code through the NPM
module names. This becomes relevant if we start moving code around in the NPM
artifact and set up package exports. e.g. our bundles might reside in `/bundles`
but we can have a mapping that makes `@angular/dev-infra/<..>` resolve to the
bundles directory (as if it would not exist) -- this is necessary for the build-worker
and commit-message-filter script resolution in the `dev-infra` repository then.