feat(ng-dev): expose entry-point with code-splitting for runtime consumers
We recently moved all `dependencies` that we bundle into the ng-dev CLI
to the `devDependencies`. This is expected but breaks consumers deeply
importing from `ng-dev` files that do not benefit from the bundling.

We should remove all these deep files and just expose public entry-points
with code splitting. This fixes the dependency issue and also allows us
to limit access to deep code. We would rather want to expose necessary
utilities explicitly, _knowing_ that they are used in consumer projects.