build: fix packaging of dev-infra package
Fixes a couple of typos, cleans up the stamping private entry-point
and adds a trick to make the release checks in COMP work. It looks
like they did not work for quite a while because the `ReleaseAction`
instance being patched as a trick is not the one used by the CLI
used for publishing the release.