build: move runtime ng-dev format deps to `dependencies`
With the recent cleanup of all dependencies and dev dependencies, we
no longer require unnecessary dependencies in dev-infra consumer projects.

As part of this change we accidentally moved two runtime ng-dev dependencies
to the `devDependencies`. We should keep buildifier and clang-format as
normal dependencies so that `yarn ng-dev format` continues to work. We also
ship prettier directly so that would be more consistent.