build: move internal dependencies from `dependencies` to `devDependencies`
There are a couple more internal dependencies that should be moved from
`dependencies` to `devDependencies`. Adds a readme for the browser update
script to make this more obvious in the future, and also moves the proto
file generation out of `bazel/` to `ng-dev/utils` as this is where the
proto-generated files are consumed. The current way of shipping the proto
file is very magical due to transitive "collecting" by the `ts_library`
rules. We want to have well-defined rules for the `bazel/` folder.