build: follow-up cleanups for rules_nodejs v5 major update
This commit contains follow-up changs for the `rules_nodejs` v5
update. It contains the following things:

* We disable `exports_directories_only` because it does not work with `ts_library`,
nor with Windows and we need a better plan going forward. This is a follow-up.

* We use the `register_toolchains` API instead of the legacy `node_repositories`. The new API
is more intuitive and helpful when running multiple Node versions.

* Updates the `extract_js_module_output.bzl` rule to handle the new linker mapping provider
API. The API changed to use depsets for much faster transitive merging. Super awesome!