fix(ng-dev/release): only match on a single character prefixed version for snapshotting (#449)
* fix(ng-dev/release): only match on a single character prefixed version for snapshotting

Previously we matched on all characters before our semver matching in the glob used for finding
the latest version for the branch we are on.  However when the branch contains something like
zonejs-1.2.3, the * matcher would capture it, leading to an unexpected syntax.  Instead we should
allow for only a single letter, typically 'v', to prefix our version tags.

* fixup! fix(ng-dev/release): only match on a single character prefixed version for snapshotting