feat(bazel/api-golden): allow for module types to be provided
Currently the API golden bazel test rule will just add `@types/<..>`
modules to the TSConfig `types` section. This is unreliable and relies
on the Bazel NodeJS linker which does not work well on Windows.

This commit refactors the logic to rely on path mappings for type
resolution inside the API golden Bazel-based tests. This guarantees
a working rule on Windows and also supports scoped packages. Currently
scoped package types do not work, like `@types/babel__core` which is
actually mapping internally to `@babel/core`.

Our trick will be to use path mappings as mentioned before, manually
computing the scoped package types (like with `@babel/core`). Additionally
referenced types will still be loaded directly as part of the compilation
inputs to support global type definitions like `NodeJS.ProcessEnv`.

The solution is not 100% ideal but it's less breaky than how type resolution
works within `@bazel/typescript` currently.. The biggest issue is that we
assume for global types that the file is called `index.d.ts`. We can allow
this to be provided in the future though, if something comes up

d