refactor(ng-dev): improve stamping for snapshot releases
Improves stamping for snapshot releases, to help with the Material
automatic docs deployment. Relying on the tags is a little problematic
since the CI version bump commit lands in the publish branch before the
release is cut in the Github UI/when the tag is created upstream.

We should rely on the workspace version similar for the actual release
mode. This is more predictable and future-proof, and solves the
automatic deployment (or other issues? in the future?) nicely.