fix(ng-dev): make `new-main-branch` command more flexible to local setup
After we will perform the default branch rename, there are three
scenarios for a contributor:

1. They manually fetch `upstream/main`, run `yarn` and then the
   migration tool.
2. They will not sync with upstream (as `upstream/master` does not
   exist), run `yarn` and then the migration tool.
3. They will manually update everything locally and in their forks. This
bullet-point is just for completion (not something we bother about here).

For (1): The `master` branch and `main` branch will locally exist. We
should ask to delete the `main` branch switch back to `master`. Then continue
with the full migration (with all other changes to the local git config)

For (2): We should run the migration as is (e.g. fetch main and set it
up locally).