build: update renovate to exclude `protobufjs` CLI deps from being updated
We have some trickery in the Protobuf Bazel macro where we need to pull
some of the `protobufjs/cli` dependencies are direct dependencies. These
dependencies need to satsify the `devDependencies` ranges of
`protobufjs`. Renovate does not know about this and attempts to update
versions we should not update/cannot update currently.

Most of the CLI packages from protobufjs/cli match with the latest ones,
so Renovate can continue to run. There are some which are using older
versions and cannot be updated. This is `escodegen` and `espree`.

https://github.com/protobufjs/protobuf.js/blob/v6.11.3/package.json#L93-L104