feat(ng-dev/pr): allow for manual override of targeted branches
Using the `manual-branch-targeting` user's can choose to manually select the
branches they would like to target rather than relying on the branches defined
by the label.  This allows for one-off manual overrides of the logic we rely on.
As noted in the warning at runtime, this does remove the attemps we make to
guarantee that all commits are valid for the targeted branches.