refactor(ng-dev/utils): update spinner to have success and failure messages, better interop with logging
Update the Spinner class to have class methods for success and failure message, automatically placing a
checkmark or x as appropriate for the method.

Additionally, updates the shared logging code to properly print a message while a Spinner is active, using
method on the spinner directly to print correctly instep with the spinner.