refactor(ng-dev): support for `stdin` text to be provided to async spawn processes
Similar to the built-in helper for `spawnSync`, the async spawn wrapper
we have should support for stdin text to be provided.

This will be used for our pre-check command where we want to transport
JSON from the release tool to the command running in older version
branches. Stdin is in such cases more safe compared to using command
line arguments (because these have max length limits)