refactor(ng-dev): do not pass `SemVer` object to precheck function to avoid version mismatches
The `SemVer` class is bundled in `ng-dev`, like other dependencies. We
should not pass an object instance of `SemVer` to the user-customizable
precheck function as it would be easy for them to run into
version-mismatch issues because SemVer utilities rely heavily on
`instanceof` checks.

e.g. in the COMP repo, we use `semver.compare` and passing the version
from bundled ng-dev in there, results in a runtime exception.

We should rather pass the string and expect users to parse it themselves
in their own user-space.