build: ensure browser `update-script` can be run as ESM
Since `nodejs_binary` is not fully ESM compatible yet, we run ESBuild
in between, creating an IIFE bundle. Long-term we could remove this when
`nodejs_binary` supports ESM with the patched module resolution, or when
we use e.g. use the new `rules_js`.

The idea of bundling this script using esbuild is not too uncommon
though, so I'm avoiding adding a comment/TODO that might stay there
for quite some time..