refactor: split ng-dev config and caching into separate file
Splits the ng-dev config cache into a separate file so that it can be
easily imported by our repo-wide jasmine macro as bootstrap script.

We wouldn't want to import/build all of `ng-dev` for every jasmine
target. It's still debatable whether we actually want to clean the
ng-dev config for all jasmine test targets, but at the same time with
this change/split, it is low-effort and ensures hermeticity/stabibilty.