fix(ng-dev/release): always publish experimental packages for lts patch
As part of a past refactoring where we made it possible for the release
tool to know whether a package is supposed to be experimental or not, we
also added a change that intended to skip experimental packages from
being published for LTS patch releases.

The idea here was that LTS experimental packages should never need to be
re-published because fixes are supposed to be only made to
non-experimental packages governed by the LTS contract. This should
technically be the case, but in practice with the way we manage our
dependencies/peer dependencies, when we bump a version for an LTS, the
experimental packages are expected to be updated as well.

To simplify this, as it is low-effort, we also publish LTS experimental
packages now.