refactor: reference other scripts through relative paths
In order to resolve the build worker script, or the commit message
filter script we should simply compute an absolute path relative
to the current script path. This is better compared to using the CJS
require resolution (going away anyway at some point with ESM), because
it makes our `dev-infra/npm_package` work without necessarily being
installed inside the `node_modules` directory. This allows us to
conveniently run the tool inside the dev-infra repo (dogfooding) and
using it for testing from `HEAD` in other repositories (without risking
to have a mismatch from the dev-infra version actually in the node
modules).