fix(github-actions): add logging and set undefined for `repoSha` and `repoFileContent`
Adds logging to aide in diagnosing issues.

Additionally, when declaring the `repoSha` and `repoFileContent` values, set the default value explicitly as `undefined`.
When not initialized as `undefined`, the only set of the value is done without a `const` or `let` which allows the value
to escape the block it is defined in.  This results in the `sha` value in the API call to be incorrect because it uses a
previous files `sha` value, or the file content as well.