feat(bazel): expose individual `.map` files for optimized bundles
Currently we do not declare individual file targets/pre-declared outputs
for map files, but we do for `.js` files. This is inconsistent and the
individual `map` files are useful to be "reference-able" when e.g. a
JS mapping size test is configured.

To achieve this we introduce a new rule for filtering outputs. This
rule will become helpful in the future as we often have scenarios where
`$(rootpath X)` cannot be used due to multiple outputs, even though e.g.
a genrule might only be interested in a single file for constructing a
command.