fix(ng-dev): do not throw if there is no diff between `g3` and `main`
It seems like multimatch previously did not throw when it received an
empty file path. We should never need to call multimatch with an empty
path, avoiding the runtime error. To fix this, we just do not
incorrectly try to derive diff information from an empty line if there
is no diff (this was a brittle logic).