refactor(ng-dev): clean-up merge tool error handling in preparation for fixes
Cleans up the merge tool error handling by not passing through failures
and having uncaught errors at the same time. This results in confusion
and easily results in errors accidentally not being caught/handled
correctly.

Instead, we always use error try/catch and throwing similar to how its
done in the release tool. This is more safe, predictable and readable.

This is in prepration of fixing some of the force skip behavior of
the merge tool. This does currently not seem to work and broke at
some point..