fix(ng-dev): force prompt should only ignore the failed validation (#730)
Currently when a validation check fails, the caretaker is prompted to
continue forcibly if the failure was non-fatal.

This seemed nice at first glance, but there is a real issue here. The
force merge run might hide actual validations which might have not been
hit before.

To fix this, we introduce an incremental validation system where we run
validations from a config (all by default) and upon failure, if
non-fatal the individual validation can be ignored, re-attempting the
merge.

Also as part of this the validation logic has been refactored/reworked
to individual files with some more complicated wrapper logic that
ensures we properly skip and report validation failures in a safe way.

PR Close #730