feat(ng-dev/pr): check for branch conflicts before confirming merge in pr merge tooling (#733)
Rather than prompting for confirmation before checking for merge conflicts, allowing
for the tool to potentially fail before the merge occurs but after the user has confirmed
the action via prompt.  The conflicts should be detected so that when prompting for final
confirmation, the user can be assured the actual merge will occur as expected.

PR Close #733