feat(ng-dev): enable the use of ng-dev credential service for pr merging (#758)
Allow for the usage of the ng-dev credential service for pr merging.  This is currently
hidden from users as we test it, though users are free to use it if they want to.  As of
now using the service for a merge requires the usage of the `--use-auth-service` flag.

PR Close #758