fix(ng-dev): always include the github token option on merge. (#771)
During the transition from using the github token to using the ng-dev auth service, we
need to always include the github token option. If we fail to include the option, when
the `useNgDevAuthService` config value is not set to `true` no authentication mechanism
for Github will be present.

PR Close #771