feat(github-actions): implement GitHub Action for creating a PR from local changes
Implement a new GitHub Action, `create-pr-for-changes` that can check
for changes in the working copy and create a pull request. It can skip
creating a pull request if there are no local changes or if a pull
request already exists for the same changes and branches (based on some
heuristic - see `github-actions/create-pr-for-changes/lib/main.ts` for
details).

This action can be useful for periodically updating checked-in files
that are auto-generated based on data from external sources (such as
events stored in a database).