fix(ng-dev): order results of forks when searching for user owned forks (#796)
It appears that Github has made a change, or has introduced a bug in which unsorted results
of forks for a repository with OWNER affiliation are returned as empty.  Once ordering is
provided the results show again as expected.  While this should not be necessary, it is an
innocuous change for us that is fine to remain indefinitely.

PR Close #796