feat(github-actions): allow specifying a base branch in `create-pr-for-changes`
Previously, the `create-pr-for-changes` GitHub action would use the
branch that the workflow was triggered on as the base branch. This
prevented the action from being used on other branches as part of a
[scheduled workflow][1], since scheduled workflows are only triggered on
the default branch of a repository.

This commit adds the ability to specify a different base branch. If not
specified, this defaults to the branch the workflow was triggered on, as
before.

[1]: https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#schedule