build: refactor rebase command to not run for pull requests
This rebase command should be a noop if no pull request is associated,
similar to what is currently done in the CLI repository.

This will help making the behavior of the command a little more obvious
compared to unknown behavior when e.g. the command would run in upstream
branches.