fix(circleci-orb): rebase script not working on windows
Fixes that the rebase script is not working on Windows by
using `spawnSync` with `shell: true`. Like we do in the ng-dev
git client (which we cannot use here).

Also fixes that the rebease command fails if there is no user
configured. We always set a user as part of rebasing. The user
may be overriden later by other scripts- like the publish snapshots
script.