fix(ng-dev): update `caretaker check` to support canonical config for g3-sync
Since we no longer will use the old legacy GitHub robot- the g3 sync
configuration is mostly used in the GitHub action for the Google
Internal test enforcement.

This means that the command in `caretaker check` needs to be updated.
While being at it, the g3sync configuration is moved directly into a
dedicated JSONC file so that both ng-dev and the github-action can
access it with shared code (yay). Also a dedicated file will make it
easier to e.g. sync the internal Copybara config to GitHub
automatically.