fix(github-actions): improve error logging and fix ESM node resolution
We recently prioritized `module` over `main` but this results in browser
code being prioritized sometimes. We revert that change and keep the
original/default ESBuild main field order.

This unveils an issue with `jsonc-parser` though- which ships an UMD
file for `main` that cannot be bundled. We can work around this by
processing `jsonc-parser`'s ESM output directly and then using that
instead of leaving resolution to ESBuild for this package.

There is an issue for this being tracked: https://github.com/microsoft/node-jsonc-parser/issues/57