fix(shared-scripts): cut down on dependencies from `@angular-devkit/build-angular`
The devkit brings in a lot of transitive dependencies and files,
potentially pushing us over the RBE input file limit

As seen here:
https://app.circleci.com/pipelines/github/angular/angular/52508/workflows/a8803c49-42cc-455e-b06a-803f6ef2edca/jobs/1246870

We can reduce this by fine-grain selecting the files we need and
managing its dependencies. These are only two dependencies and it's
acceptable taking responsibility for it- especially since we have
test coverage for it.