fix(github-actions): use proper variable names when accessing deploy metadata
The wrong variables were used and the GitHub action did not even fail,
but instead use some undefined value, causing the Firebase URLs to
include some non-deterministic string.

Also the comment did not get posted because the PR number was not a
valid one. Even here, the action did not print a failure..