perf(github-actions): compress preview artifacts before uploading
The `actions/upload` infrastructure will upload all files individually,
but we can compress it directly beforehand and just upload a single
smaller file.

This means that we need to unzip -> and then untar as GitHub
always zips up the artifacts regardless when they are uploaded.

This is expected and recommended as per:
https://github.com/actions/upload-artifact