fix(bazel): downlevel async generators along with async/await
This is necessary as ESBuild does not support it. It is also
needed to bring our pipeline in sync with the Angular CLI.

Note that I combined both linker & optimization into a single
plugin so that we do not have to run slow Babel compilations
more than once. This matches with the Angular CLI too.

Long-term we might be able to leverage a plugin from them
directly when the ESBuild builder is the default.