test: update size golden after `isSideEffectFree` fix
The size test changed when we fixed `isSideEffectFree`. Note that
I investigated why there is a minimal increase even though we should
treat more files as side-effect-free. The reason is that due to the
pure annotations Terser optimizes slightly different. In larger apps,
this would be a size reduction but here due to slightly differently
optimized code (i.e. some functions are inlined / some not)- it causes
a small increase.

e.g. rxjs `createErrorClass` is only used once but now is no longer
inlined, hence having a `function` modifier compared to it previously
being inlined without the function declaration overhead.