feat(bazel): always consider transitive specs for spec-entrypoint
We recently switched all from transitive to just direct. We should
keep this for the bootstrap code as there is a legitimate ordering
issue if a bootstrap file depends on another and they would then be
brought into with an unexpected order.

For specs we should consider all transitive ones too, just to be on the
safe side of never missing any specs.