feat(ng-dev): support exceptional minor release trains
An exceptional minor may be cut when work already started on
a major release train. The major could still be in the `main`
branch, or already have moved into a dedicated version branch
as part of feature-freeze or release-candidate.

In such cases, if the Angular team suddenly decides that another
minor for the current "latest" major is needed. e.g. to account for
a TypeScript version range expansion- then a new version branch needs
to be created. This branch *CANNOT* be based on `main` or the rc/FF
branch because major features/breaking-changes might have landed
already.

To support this use-case though (named "exceptional minors"), we will
allow branching-off from the current patch branch. The minor should
not be directly relesed as there needs to be time to e.g. backport
minor changes already landed in `main` / or the new work needs to be
performed (e.g. the TS update). This means that the newly-created branch
will *NOT* directly become the new patch branch. Instead, the new branch
will operate similar to the feature-freeze/RC branches, allowing for
pre-releases. This is the new exceptional-minor release train.

As a workaround in the past we abused the FF/RC release train for this,
but this will not work when the major has already branched off into
RC/FF. To support this, a separate new release-train is needed.