fix(ng-dev): properly catch target label/branch errors
Currently not having a target label, or other branch determination
errors by the target label configs- results in an error being raised.

These errors are not handled properly and their full object is printed.
This is rather confusing and the errors should be handled in a way
that match other merge tool errors.