feat(ng-dev): support exceptional minors in release publish tool
Introduces three new actions:

* An action for initiating an exceptional minor. This is the point where
  we branch-off from the existing patch.
* An action for cutting pre-releases of an exceptional minor. e.g.
  bumping next.0 to next.1, or rc.0 to rc.1
* An action for cutting the first RC in an exceptional minor.

Also we update the existing actions:

* Cut Stable: It will now take the exceptional minor, if there is one.
  Allowing an exceptional minor to becomethe new "patch".

The base logic for how pre-releases and the first RC is cut has been
a little more updated so that more code duplication can be avoided.

Really a first RC action is an extension of the normal pre-releases
action. This is now achieved by extending from it and overidding where
needed.