feat(ng-dev): support exceptional minor release train in merge targeting
We are still exploring an alternative long-term solution with
a potentially simpler mental model, but for now we are going
with a minimal-invasive solution that allows for exceptional minors
to become an offically-supported situation.

If there is no exceptional minor, target: minor will behave like before
If there is an exceptional minor, target: patch always goes in there
If there is an exceptional minor, target: minor will only behave
  differently if the GitHub PR target is explicitly set to the exceptional minor.

So in TL:DR: Everything remains the same unless there is an exceptional
minor. But even then, targeting like before would work the same without
any required process/model change.