feat(ng-dev): delete the temporary exceptional minor NPM dist tag when turning stable
When an exceptional minor turns "stable" and becomes the new patch, the
temporary NPM dist tag can be deleted. We do this to avoid having
unnecessary NPM dist tags show up on NPM.

The whole tag is only needed for being able to publish and not
overriding "latest". `npm publish` always requires a tag, and falls
back to overriding "latest" if no expliit tag is provided.