build: fix scorecard signing by ensuring permissions are only set for job (#964)
First of all, scorecard does not print a useful error at all. Looking
through issues and their verify workflow logic, it seems like the
permissions for the scorecard action workflow *need* to be set on the
actual job itself, while the workflow-level permissions *cannot* contain
any `write` permissions.

We already got further with the errors by enabling id-token write
properly, but now the signing fails with an unknown error that we
attempt to fix using this commit.