fix(github-actions): branch manager incorrectly syncs all PRs
The branch manager logic for syncing merge ready PRs is flawed
and currently always causes all PRs to be synced. This can be
quite a lot, especially in repos like FW.

This caused queueing of GitHub actions in the dev-infra repo
and slowed down updates of the mergeability status.

Related to: #998