feat(ng-dev): support autosquashing fixup commits for the API merge strategy
Add support for autosquashing fixup commits before using the Github API to merge pull requests
with the API merge strategy, for non-squash usage.

The pull request is updated with the rebased commits and then immediately merged as the statuses can
be safely skipped as they will match the previous run with only an autosquash changing.