feat(ng-dev/release): properly mark releases as latest on GitHub
GitHub currently will always mark the most recent release as "latest"
and show it on the repository.

This required us to release in a certain order. This commit fixes this
by only marking a release as "latest" based on the active release trains
and chosen release action. This allows us to release in any order, and
e.g. prevents LTS releases from accidentally showing up as "latest" on
GitHub.

Also cleaned up some boolean flags in tests to make tests much more
readable.