fix(bazel): allow lookup failures when using true-case-path
There are some cases it seems where certain symbols (potentially ~) cause inconsistencies in true-case-paths
ability to find the path. As this project is abandoned, we do not expect any sort of fix in the
future. This can be safely done as if the file actually does not exist, then a failure will
occur when we actually try to act on the file, rather than this early failure.