build: fix PR builds always getting cancelled
Currently when a PR lands in `main`, or when another PR is created,
they all _always_ share the same concurrency group because the
concurrency group `github.ref` is always the main branch due to
`pull_request_target` being used.

We can fix this by switching to `pull_request` which also has the
benefit of *NOT* executing user code in trusted VMs. Another benefit
is that action config changes can be directly tested in PRs