feat(github-actions): dynamically set the resources available on the runner during bazel setup
Dynamically set the RAM and CPU available on the runner in the bazelrc during bazel setup.  This allows for all OSes to
know the amount of resources available during execution, even when Bazel can't figure it out due to VM usage.  This is done
using a shell wrapped node script to make for easier multiple OS support.  The execution of the check/update is sub 50ms
during local testing and does not present much additional overhead and should be worth the cost.