feat(api-gen): switch from nunjucks to tsx
Based on top of #1387

This changes rendering from nunjucks to TSX with preact. Benefits:
* Full type safety in templates
* preact + tsx is well-supported and widely used
* TSX is much more familiar for a much wider set of developers

Note that the tsconfig change here is not the most up-to-date recommendation for TSX with preact. For some reason, trying to specify `tsx: 'react-jsx'` doesn't seem to carry through the way it's supposed to and `tsc` always ends up receiving `tsx: 'react'`. This older configuration still works fine, though.