fix(api-gen): support multiple collections for one module in manifest (#1487)
Currently the manifest generation assumes each module has exactly one
API collection. This change makes it so that a module may be split up
into multiple collections, all of which will be unioned together into
the manifest. The impetus for this change is adding hand-written
collections for Angular element and block APIs.

PR Close #1487