refactor(api-gen): remove cross-linking
As authored, the rendering code added links to text inside inline markdown code blocks if the content of the code exactly matched a doc entity name. This is not viable in the actual rule because, at time of rendering, it does not have knowledge of the entire corpus of entities (only those that were passed to the bazel rule). This commit completely removes this behavior.