feat(ng-dev): support enforcing certain checks to be running
This can be used in the future to enforce the unified status check, or
to enforce additional thing. Right now it can be used to enforce some
checks that aren't guaranteed to run otherwise.

In the future we may remove the configuration option and always default
to the unified status check. This is the goundwork for future changes.