fix(bazel): support class member entry names with no name for signatures
Signatures might be represented for now using a member entry, using
a name of an empty string. This unveiled a bug in the ToC grouping
where multiple overloads of the same member name (or empty name) were
silently skipped due to a truthy check that should instead check
explicitly for `undefined`, following strict null checks semantics.