ci: use keepUpdatedLabel to control autorebasing of renovate PRs
Using keepUpdatedLabel we can turn on and off the autorebasing functionality of renovate.  Rather than having renovate always
autorebase, we will enable this using a label which renovate will automatically create PRs with. This will cause the current
behavior to continue by default, but allow us to remove the label for PRs which are blocked, but renovate would automatically
regenerate against our wishes.

The label we will use in `renovate managed`.